♻️refactor: V2 패키지로 모임 목록 조회 이동 [#122]
