+++ src/main/java/team/wego/wegobackend/group/v2/application/dto/response/GetGroupV2Response.java
@@ -1,4 +1,4 @@
-package team.wego.wegobackend.group.application.dto.v2.response;
+package team.wego.wegobackend.group.v2.application.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
